extracted some cheerio related stuff
